Crop plate

Crops plate into a grid

crop_plate(img, points, cols = 4, rows = 3, prefix = "", col.resize = 1,
  row.resize = 1.2, adjust.cell = 0, return.images = FALSE)

Arguments

img
Image object. Recommended to use EBImage objects, but any 3-dimensional matrix would work

points
A list containing topleft, topright, bottomleft, and bottomright entries. Each entry must be a names numeric vector of length two with elements "m.cy" and "m.cy", which correspond to the x and y coordinates of each point

cols, rows
Number of columns and rows desired in the grid.

prefix
String with file prefix to append to filenames of resulting grid files.

col.resize, row.resize
NOT IMPLEMENTED. Eventually allowing to

adjust.cell
Number of pixels to increase every side of each cell rectangle.

return.images
Logical indicating whether image objects should be returned for the grid. Defaults to FALSE, which only generates output files.