Get images and masks

Makes a list of matching images and their mask. Useful to bulk display mask images

get_images_and_masks(img_dir, mask_dir)

Arguments

img_dir
A directory where images are located

mask_dir
A directory containing only mask files

Value

A lust where each element is a list with elements img and mask. Each of those elements is an EBImage image object.

Details

Takes two directories, one for images and another for masks, then for every file in the mask directory (mask_dir), it looks for an identically named file in the image directory (img_dir), creating a list of matching images and masks.